move PriceTextInput to its own file
