+++ src/nft/components/profile/list/NFTListRow.tsx
@@ -8,11 +8,14 @@ import { body, bodySmall, subhead } from 'nft/css/common.css'
 import { useSellAsset } from 'nft/hooks'
 import { ListingMarket, ListingWarning, WalletAsset } from 'nft/types'
 import { LOOKS_RARE_CREATOR_BASIS_POINTS } from 'nft/utils'
+import { formatEth, formatUsdPrice } from 'nft/utils/currency'
+import { fetchPrice } from 'nft/utils/fetchPrice'
 import { Dispatch, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components/macro'
 
 import * as styles from './ListPage.css'
-import { EthPriceDisplay, PriceTextInput, SetPriceMethod } from './NFTListingsGrid'
+import { SetPriceMethod } from './NFTListingsGrid'
+import { PriceTextInput } from './PriceTextInput'
 import { RoyaltyTooltip } from './RoyaltyTooltip'
 
 interface NFTListRowProps {
@@ -361,5 +364,35 @@ const MarketplaceRow = ({
         </Column>
       </Column>
     </Row>
+  )
+}
+
+const EthPriceDisplay = ({ ethPrice = 0 }: { ethPrice?: number }) => {
+  const [ethConversion, setEthConversion] = useState(3000)
+  useEffect(() => {
+    fetchPrice().then((price) => {
+      setEthConversion(price ?? 0)
+    })
+  }, [])
+
+  return (
+    <Column width="full">
+      <Row width="full" justifyContent="flex-end" color={ethPrice !== 0 ? 'textPrimary' : 'textSecondary'}>
+        {ethPrice !== 0 ? (
+          <>
+            <Column>
+              <Box className={body} color="textPrimary" textAlign="right" marginLeft="12" marginRight="0">
+                {formatEth(ethPrice)} ETH
+              </Box>
+              <Box className={body} color="textSecondary" textAlign="right" marginLeft="12" marginRight="0">
+                {formatUsdPrice(ethPrice * ethConversion)}
+              </Box>
+            </Column>
+          </>
+        ) : (
+          '- ETH'
+        )}
+      </Row>
+    </Column>
   )
 }
